requirements.txt
python-keystoneclient
openstacksdk
python-novaclient
django
